% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_analysis.R
\name{analyze_patterns_multi}
\alias{analyze_patterns_multi}
\title{Multi-Group Pattern Analysis (Alias for analyze_patterns)}
\usage{
analyze_patterns_multi(...)
}
\arguments{
\item{...}{All arguments passed to analyze_patterns()}
}
\value{
Same as analyze_patterns()
}
\description{
This function is now an alias for analyze_patterns() which auto-detects
the number of groups. For multi-group analysis, use analyze_patterns() directly.
}
