% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_analysis.R
\name{analyze_specific_pattern}
\alias{analyze_specific_pattern}
\title{Analyze Specific Pattern (Wildcard Aware)}
\usage{
analyze_specific_pattern(seq_strings, pattern, n_seqs, n_states, correction)
}
\description{
Analyze Specific Pattern (Wildcard Aware)
}
\keyword{internal}
