% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_patterns.R
\name{apply_filters_optimized}
\alias{apply_filters_optimized}
\title{Apply filters efficiently}
\usage{
apply_filters_optimized(
  df,
  min_support,
  min_count,
  start_state,
  end_state,
  contains_state
)
}
\description{
Apply filters efficiently
}
\keyword{internal}
