% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_rules.R
\name{build_conditional_fp_tree}
\alias{build_conditional_fp_tree}
\title{Build Conditional FP-Tree}
\usage{
build_conditional_fp_tree(conditional_patterns, min_support, n_transactions)
}
\arguments{
\item{conditional_patterns}{List of conditional patterns}

\item{min_support}{Minimum support threshold}
}
\value{
Conditional FP-tree
}
\description{
Build Conditional FP-Tree
}
