% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_rules.R
\name{build_fp_tree}
\alias{build_fp_tree}
\title{Build FP-Tree Structure}
\usage{
build_fp_tree(transactions, min_support, verbose = TRUE)
}
\arguments{
\item{transactions}{List of transactions}

\item{min_support}{Minimum support threshold}

\item{verbose}{Whether to show progress}
}
\value{
FP-tree structure
}
\description{
Build FP-Tree Structure
}
