% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_utils.R
\name{calculate_itemset_support}
\alias{calculate_itemset_support}
\title{Calculate Support for Itemset}
\usage{
calculate_itemset_support(itemset, trans_matrix)
}
\arguments{
\item{itemset}{Vector of items}

\item{trans_matrix}{Binary transaction matrix}
}
\value{
Support value (proportion of transactions containing the itemset)
}
\description{
Computes the support (frequency) of an itemset in the transaction data.
}
