% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_comparison.R
\name{compare_sequences_multi}
\alias{compare_sequences_multi}
\title{Multi-Group Sequence Comparison (Alias for compare_sequences)}
\usage{
compare_sequences_multi(...)
}
\arguments{
\item{...}{All arguments passed to compare_sequences()}
}
\value{
Same as compare_sequences()
}
\description{
This function is now an alias for compare_sequences() which auto-detects
the number of groups. For multi-group analysis, use compare_sequences() directly.
}
