% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_indices.R
\name{compute_change_volatility}
\alias{compute_change_volatility}
\title{Compute change volatility - how much the system changes}
\usage{
compute_change_volatility(states)
}
\description{
Compute change volatility - how much the system changes
}
