% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_indices.R
\name{compute_comprehensive_durations}
\alias{compute_comprehensive_durations}
\title{Compute comprehensive durations considering all possible states}
\usage{
compute_comprehensive_durations(
  observed_states,
  observed_durations,
  all_possible_states
)
}
\description{
Compute comprehensive durations considering all possible states
}
