% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_indices.R
\name{compute_concentration_index}
\alias{compute_concentration_index}
\title{Compute concentration index (similar to Gini coefficient)}
\usage{
compute_concentration_index(frequencies)
}
\description{
Compute concentration index (similar to Gini coefficient)
}
