% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_methods.R
\name{compute_dl_distance}
\alias{compute_dl_distance}
\title{Compute Damerau-Levenshtein Distance}
\usage{
compute_dl_distance(data)
}
\arguments{
\item{data}{A data frame or matrix where rows are sequences and columns are time points}
}
\value{
A distance object of class 'dist'
}
\description{
Compute Damerau-Levenshtein Distance
}
\keyword{internal}
