% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_indices.R
\name{compute_emergent_state}
\alias{compute_emergent_state}
\title{Compute emergent state - only if it becomes MORE dominant than the attractor}
\usage{
compute_emergent_state(states)
}
\description{
Compute emergent state - only if it becomes MORE dominant than the attractor
}
