% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_indices.R
\name{compute_favorable_state_measures}
\alias{compute_favorable_state_measures}
\title{Compute favorable state measures}
\usage{
compute_favorable_state_measures(states, favorable_states)
}
\description{
Compute favorable state measures
}
