% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_indices.R
\name{compute_feedback_loops}
\alias{compute_feedback_loops}
\title{Compute feedback loop measures}
\usage{
compute_feedback_loops(states, transition_matrix)
}
\description{
Compute feedback loop measures
}
