% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_indices.R
\name{compute_initial_state_influence}
\alias{compute_initial_state_influence}
\title{Compute initial state influence measures}
\usage{
compute_initial_state_influence(states)
}
\description{
Compute initial state influence measures
}
