% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_methods.R
\name{compute_jw_distance}
\alias{compute_jw_distance}
\title{Compute Jaro-Winkler Distance}
\usage{
compute_jw_distance(data, p = 0.1)
}
\arguments{
\item{data}{A data frame or matrix where rows are sequences and columns are time points}

\item{p}{Prefix scaling factor (default: 0.1)}
}
\value{
A distance object of class 'dist'
}
\description{
Compute Jaro-Winkler Distance
}
\keyword{internal}
