% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_methods.R
\name{compute_lcs_stringdist_distance}
\alias{compute_lcs_stringdist_distance}
\title{Compute LCS Distance via stringdist}
\usage{
compute_lcs_stringdist_distance(data)
}
\arguments{
\item{data}{A data frame or matrix where rows are sequences and columns are time points}
}
\value{
A distance object of class 'dist'
}
\description{
Compute LCS Distance via stringdist
}
\keyword{internal}
