% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_indices.R
\name{compute_markov_parameters}
\alias{compute_markov_parameters}
\title{Compute Markov chain parameters}
\usage{
compute_markov_parameters(states, transition_matrix)
}
\description{
Compute Markov chain parameters
}
