% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_methods.R
\name{compute_osa_distance}
\alias{compute_osa_distance}
\title{Compute OSA (Optimal String Alignment) Distance}
\usage{
compute_osa_distance(data)
}
\arguments{
\item{data}{A data frame or matrix where rows are sequences and columns are time points}
}
\value{
A distance object of class 'dist'
}
\description{
Compute OSA (Optimal String Alignment) Distance
}
\keyword{internal}
