% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_methods.R
\name{compute_qgram_distance}
\alias{compute_qgram_distance}
\title{Compute Q-gram Distance}
\usage{
compute_qgram_distance(data, q = 2)
}
\arguments{
\item{data}{A data frame or matrix where rows are sequences and columns are time points}

\item{q}{Q-gram size (default: 2)}
}
\value{
A distance object of class 'dist'
}
\description{
Compute Q-gram Distance
}
\keyword{internal}
