% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_methods.R
\name{compute_soundex_distance}
\alias{compute_soundex_distance}
\title{Compute Soundex Distance}
\usage{
compute_soundex_distance(data)
}
\arguments{
\item{data}{A data frame or matrix where rows are sequences and columns are time points}
}
\value{
A distance object of class 'dist'
}
\description{
Compute Soundex Distance
}
\keyword{internal}
