% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_indices.R
\name{compute_state_spells}
\alias{compute_state_spells}
\title{Compute spells for a specific state}
\usage{
compute_state_spells(states, target_state)
}
\description{
Compute spells for a specific state
}
