% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_indices.R
\name{compute_state_type_stability}
\alias{compute_state_type_stability}
\title{Compute stability for favorable or unfavorable states}
\usage{
compute_state_type_stability(states, favorable_states, is_favorable_type)
}
\description{
Compute stability for favorable or unfavorable states
}
