% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_indices.R
\name{compute_system_stability}
\alias{compute_system_stability}
\title{Compute system stability - how stable the system is overall}
\usage{
compute_system_stability(states)
}
\description{
Compute system stability - how stable the system is overall
}
