% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_indices.R
\name{compute_transition_complexity}
\alias{compute_transition_complexity}
\title{Compute transition complexity considering favorable/unfavorable states}
\usage{
compute_transition_complexity(
  states,
  transition_matrix,
  favorable_states = NULL
)
}
\description{
Compute transition complexity considering favorable/unfavorable states
}
