% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_motifs.R
\name{compute_type_transitions}
\alias{compute_type_transitions}
\title{Compute type-to-type transitions}
\usage{
compute_type_transitions(
  type_sequences,
  type_names,
  n_sequences,
  test_significance,
  correction,
  alpha
)
}
\description{
Compute type-to-type transitions
}
\keyword{internal}
