% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_rules.R
\name{create_empty_rules_object}
\alias{create_empty_rules_object}
\title{Create Empty Rules Object}
\usage{
create_empty_rules_object(algorithm)
}
\arguments{
\item{algorithm}{Algorithm name}
}
\value{
Empty association rules object
}
\description{
Create Empty Rules Object
}
