% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_tna_support.R
\name{create_mock_group_tna}
\alias{create_mock_group_tna}
\title{Create a mock group_tna object for testing}
\usage{
create_mock_group_tna(
  groups = c("GroupA", "GroupB", "GroupC"),
  n_sequences = 5,
  n_timepoints = 4,
  states = c("Active", "Average", "Disengaged")
)
}
\arguments{
\item{groups}{List of group names}

\item{n_sequences}{Number of sequences per group}

\item{n_timepoints}{Number of time points}

\item{states}{Possible states}
}
\value{
Mock group_tna object
}
\description{
Create a mock group_tna object for testing
}
