% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_motifs.R
\name{discover_gapped_patterns}
\alias{discover_gapped_patterns}
\title{Discover gapped patterns automatically}
\usage{
discover_gapped_patterns(
  sequences,
  all_states,
  min_gap,
  max_gap,
  n_sequences,
  n_states,
  min_support,
  min_count,
  test_significance,
  correction,
  alpha,
  verbose
)
}
\description{
Discover gapped patterns automatically
}
\keyword{internal}
