% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_motifs.R
\name{discover_meta_paths}
\alias{discover_meta_paths}
\title{Discover meta-paths automatically}
\usage{
discover_meta_paths(
  sequences,
  type_sequences,
  type_names,
  state_to_type,
  min_length,
  max_length,
  n_sequences,
  n_states,
  min_support,
  min_count,
  test_significance,
  correction,
  alpha,
  verbose
)
}
\description{
Discover meta-paths automatically
}
\keyword{internal}
