% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_methods.R
\name{.create_stringdist_matrix}
\alias{.create_stringdist_matrix}
\title{Create pairwise distance matrix using stringdist methods}
\usage{
.create_stringdist_matrix(sequences, method, ...)
}
\arguments{
\item{sequences}{Character vector of sequences}

\item{method}{StringDist method name}

\item{...}{Additional parameters for stringdist}
}
\value{
A distance object of class 'dist'
}
\description{
Create pairwise distance matrix using stringdist methods
}
\keyword{internal}
