% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_patterns.R
\name{extract_abstract_patterns}
\alias{extract_abstract_patterns}
\title{Extract abstract patterns}
\usage{
extract_abstract_patterns(
  seq_data,
  min_gap,
  max_gap,
  min_support,
  min_count,
  verbose
)
}
\description{
Extract abstract patterns
}
\keyword{internal}
