% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_patterns.R
\name{extract_full_optimized}
\alias{extract_full_optimized}
\title{Optimized full sequence analysis}
\usage{
extract_full_optimized(sequences, n_sequences, fast_mode, verbose)
}
\description{
Optimized full sequence analysis
}
\keyword{internal}
