% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_patterns.R
\name{extract_gapped_optimized}
\alias{extract_gapped_optimized}
\title{Optimized gapped pattern discovery}
\usage{
extract_gapped_optimized(
  sequences,
  all_states,
  min_gap,
  max_gap,
  n_sequences,
  min_support,
  min_count,
  fast_mode,
  verbose
)
}
\description{
Optimized gapped pattern discovery
}
\keyword{internal}
