% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_patterns.R
\name{extract_gapped_vectorized}
\alias{extract_gapped_vectorized}
\title{Extract Gapped Patterns (Vectorized)}
\usage{
extract_gapped_vectorized(seq_data, min_gap, max_gap, verbose)
}
\description{
Extract Gapped Patterns (Vectorized)
}
\keyword{internal}
