% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_tna_support.R
\name{extract_group_tna_info}
\alias{extract_group_tna_info}
\title{Extract metadata from group_tna object}
\usage{
extract_group_tna_info(group_tna_obj)
}
\arguments{
\item{group_tna_obj}{A group_tna object}
}
\value{
List with metadata information
}
\description{
Extract metadata from group_tna object
}
