% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_patterns.R
\name{filter_patterns}
\alias{filter_patterns}
\title{Filter Patterns}
\usage{
filter_patterns(x, ...)
}
\arguments{
\item{x}{An object of class \code{sequence_patterns}}

\item{...}{Additional arguments (currently unused, for future extensibility)}
}
\value{
A data frame of patterns
}
\description{
Extracts the patterns data frame from a sequence_patterns object.
}
