% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_rules.R
\name{find_frequent_itemsets_apriori}
\alias{find_frequent_itemsets_apriori}
\title{Find Frequent Itemsets using Apriori Algorithm}
\usage{
find_frequent_itemsets_apriori(
  trans_matrix,
  min_support,
  max_length,
  verbose = TRUE
)
}
\arguments{
\item{trans_matrix}{Binary transaction matrix}

\item{min_support}{Minimum support threshold}

\item{max_length}{Maximum itemset length}

\item{verbose}{Whether to show progress}
}
\value{
List of frequent itemsets
}
\description{
Find Frequent Itemsets using Apriori Algorithm
}
