% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_patterns.R
\name{find_pattern_matches_optimized}
\alias{find_pattern_matches_optimized}
\title{Optimized pattern matching}
\usage{
find_pattern_matches_optimized(seq, parts)
}
\description{
Optimized pattern matching
}
\keyword{internal}
