% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_rules.R
\name{generate_association_rules}
\alias{generate_association_rules}
\title{Generate Association Rules from Frequent Itemsets}
\usage{
generate_association_rules(
  frequent_itemsets,
  trans_matrix,
  min_confidence,
  min_lift
)
}
\arguments{
\item{frequent_itemsets}{List of frequent itemsets}

\item{trans_matrix}{Transaction matrix}

\item{min_confidence}{Minimum confidence threshold}

\item{min_lift}{Minimum lift threshold}
}
\value{
Data frame of association rules
}
\description{
Generate Association Rules from Frequent Itemsets
}
