% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_rules.R
\name{generate_conditional_patterns}
\alias{generate_conditional_patterns}
\title{Generate Conditional Pattern Base}
\usage{
generate_conditional_patterns(fp_tree, item)
}
\arguments{
\item{fp_tree}{FP-tree structure}

\item{item}{Target item}
}
\value{
List of conditional patterns
}
\description{
Generate Conditional Pattern Base
}
