% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_rules.R
\name{generate_rules_from_patterns}
\alias{generate_rules_from_patterns}
\title{Generate Rules from FP-Growth Patterns}
\usage{
generate_rules_from_patterns(
  frequent_patterns,
  trans_matrix,
  min_confidence,
  min_lift
)
}
\arguments{
\item{frequent_patterns}{List of frequent patterns}

\item{trans_matrix}{Transaction matrix}

\item{min_confidence}{Minimum confidence threshold}

\item{min_lift}{Minimum lift threshold}
}
\value{
Data frame of association rules
}
\description{
Generate Rules from FP-Growth Patterns
}
