% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_utils.R
\name{get_rules_network}
\alias{get_rules_network}
\title{Get Rules Network Data}
\usage{
get_rules_network(rules, top_n = Inf)
}
\arguments{
\item{rules}{Association rules object or data frame}

\item{top_n}{Number of top rules to include (default: Inf for all)}
}
\value{
List containing 'nodes' (data frame) and 'edges' (data frame)
}
\description{
Extracts network data (nodes and edges) from association rules.
}
