% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_rules.R
\name{head.association_rules}
\alias{head.association_rules}
\title{Head Method for Association Rules}
\usage{
\method{head}{association_rules}(x, n = 6, ...)
}
\arguments{
\item{x}{Association rules object}

\item{n}{Number of rules to show}

\item{...}{Additional arguments}
}
\description{
Head Method for Association Rules
}
