% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_rules.R
\name{insert_transaction_fp}
\alias{insert_transaction_fp}
\title{Insert Transaction into FP-Tree}
\usage{
insert_transaction_fp(fp_tree, items)
}
\arguments{
\item{fp_tree}{FP-tree structure}

\item{items}{Sorted items to insert}
}
\description{
Insert Transaction into FP-Tree
}
