% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_tna_support.R
\name{is_group_tna}
\alias{is_group_tna}
\title{Check if object is a group_tna object}
\usage{
is_group_tna(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical indicating if object is group_tna
}
\description{
Check if object is a group_tna object
}
