% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_analysis.R
\name{mine_abstract_fast}
\alias{mine_abstract_fast}
\title{Optimized Abstract Pattern Mining}
\usage{
mine_abstract_fast(
  seq_list,
  min_gap,
  max_gap,
  min_sup,
  min_cnt,
  n_seqs,
  verbose
)
}
\description{
Optimized Abstract Pattern Mining
}
\keyword{internal}
