% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_rules.R
\name{mine_fp_tree}
\alias{mine_fp_tree}
\title{Mine FP-Tree for Frequent Patterns}
\usage{
mine_fp_tree(fp_tree, min_support, n_transactions, verbose = TRUE)
}
\arguments{
\item{fp_tree}{FP-tree structure}

\item{min_support}{Minimum support threshold}

\item{verbose}{Whether to show progress}
}
\value{
List of frequent patterns
}
\description{
Mine FP-Tree for Frequent Patterns
}
