% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_analysis.R
\name{mine_full_fast}
\alias{mine_full_fast}
\title{Optimized Full Sequence Mining}
\usage{
mine_full_fast(seq_strings, min_sup, min_cnt, n_seqs, verbose)
}
\description{
Optimized Full Sequence Mining
}
\keyword{internal}
