% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_analysis.R
\name{mine_gapped_fast}
\alias{mine_gapped_fast}
\title{Optimized Gapped Mining (Vectorized Regex)}
\usage{
mine_gapped_fast(
  seq_strings,
  states,
  min_gap,
  max_gap,
  min_sup,
  min_cnt,
  n_seqs,
  verbose
)
}
\description{
Optimized Gapped Mining (Vectorized Regex)
}
\keyword{internal}
