% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_analysis.R
\name{mine_ngrams_fast}
\alias{mine_ngrams_fast}
\title{Optimized N-Gram Mining}
\usage{
mine_ngrams_fast(seq_list, min_len, max_len, min_sup, min_cnt, n_seqs, verbose)
}
\description{
Optimized N-Gram Mining
}
\keyword{internal}
