% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_viz.R
\name{plot_itemset_frequency}
\alias{plot_itemset_frequency}
\title{Plot Itemset Frequency Bar Chart}
\usage{
plot_itemset_frequency(
  rules,
  top_n = 20,
  type = "both",
  main = "Itemset Frequencies"
)
}
\arguments{
\item{rules}{Association rules object or data frame}

\item{top_n}{Number of top itemsets to show}

\item{type}{Type of itemsets to show ("antecedent", "consequent", "both")}

\item{main}{Plot title}
}
\description{
Creates a bar chart showing the frequency of frequent itemsets.
}
