% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_viz.R
\name{plot_rules_matrix}
\alias{plot_rules_matrix}
\title{Plot Rules Matrix}
\usage{
plot_rules_matrix(rules, top_n = 50, color_by = "lift", main = "Rules Matrix")
}
\arguments{
\item{rules}{Association rules object or data frame}

\item{top_n}{Number of top rules to show}

\item{color_by}{Metric to use for coloring ("support", "confidence", "lift")}

\item{main}{Plot title}
}
\description{
Creates a matrix visualization of rules showing antecedent-consequent relationships.
}
