% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_rules.R
\name{print.association_rules}
\alias{print.association_rules}
\title{Print Method for Association Rules}
\usage{
\method{print}{association_rules}(x, ...)
}
\arguments{
\item{x}{Association rules object}

\item{...}{Additional arguments}
}
\description{
Print Method for Association Rules
}
