% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_comparison.R
\name{print.compare_sequences}
\alias{print.compare_sequences}
\title{Print Method for compare_sequences Objects}
\usage{
\method{print}{compare_sequences}(x, ...)
}
\arguments{
\item{x}{A compare_sequences object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for compare_sequences Objects
}
