% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_comparison.R
\name{print.compare_sequences_multi}
\alias{print.compare_sequences_multi}
\title{Print method for compare_sequences_multi objects}
\usage{
\method{print}{compare_sequences_multi}(x, ...)
}
\arguments{
\item{x}{compare_sequences_multi object}

\item{...}{additional arguments}
}
\description{
Print method for compare_sequences_multi objects
}
