% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_tna_support.R
\name{print.group_tna}
\alias{print.group_tna}
\title{Print method for mock group_tna objects}
\usage{
\method{print}{group_tna}(x, ...)
}
\arguments{
\item{x}{Mock group_tna object}

\item{...}{Additional arguments}
}
\description{
Print method for mock group_tna objects
}
