% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_clustering_cov.R
\name{print.mmm_cov_result}
\alias{print.mmm_cov_result}
\title{Print Method for MMM Covariate Results}
\usage{
\method{print}{mmm_cov_result}(x, ...)
}
\arguments{
\item{x}{A mmm_cov_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Custom print method for MMM clustering results with covariates.
}
