% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_analysis.R
\name{print.pattern_analysis_multi}
\alias{print.pattern_analysis_multi}
\title{Print method for pattern_analysis_multi objects}
\usage{
\method{print}{pattern_analysis_multi}(x, ...)
}
\arguments{
\item{x}{pattern_analysis_multi object}

\item{...}{additional arguments}
}
\description{
Print method for pattern_analysis_multi objects
}
