% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_indices.R
\name{print_indices_summary}
\alias{print_indices_summary}
\title{Print summary of sequence indices - Complexity Systems Version}
\usage{
print_indices_summary(indices_result, top_n = 5)
}
\arguments{
\item{indices_result}{Result from compute_sequence_indices()}

\item{top_n}{Number of top sequences to show for each measure}
}
\description{
Print summary of sequence indices - Complexity Systems Version
}
