% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_utils.R
\name{print_patterns_summary}
\alias{print_patterns_summary}
\title{Print Patterns Summary}
\usage{
print_patterns_summary(patterns_df, title = "Patterns", n_show = 15)
}
\arguments{
\item{patterns_df}{Data frame of patterns}

\item{title}{Title for the section}

\item{n_show}{Number of patterns to show}
}
\description{
Print Patterns Summary
}
\keyword{internal}
