% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_utils.R
\name{rank_association_rules}
\alias{rank_association_rules}
\title{Rank Association Rules}
\usage{
rank_association_rules(rules, by = "lift", decreasing = TRUE)
}
\arguments{
\item{rules}{Data frame of association rules}

\item{by}{Metric to sort by ("support", "confidence", "lift", "conviction")}

\item{decreasing}{Whether to sort in decreasing order}
}
\value{
Sorted rules data frame
}
\description{
Sorts association rules by specified metric.
}
