% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_motifs.R
\name{search_gapped_pattern}
\alias{search_gapped_pattern}
\title{Search for specific gapped pattern}
\usage{
search_gapped_pattern(
  sequences,
  pattern,
  n_sequences,
  n_states,
  test_significance,
  correction,
  alpha
)
}
\description{
Search for specific gapped pattern
}
\keyword{internal}
