% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_motifs.R
\name{search_meta_path_schema}
\alias{search_meta_path_schema}
\title{Search for specific meta-path schema}
\usage{
search_meta_path_schema(
  sequences,
  type_sequences,
  schema,
  state_to_type,
  node_types,
  n_sequences,
  n_states,
  test_significance,
  correction,
  alpha
)
}
\description{
Search for specific meta-path schema
}
\keyword{internal}
