% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_patterns.R
\name{search_multi_wildcard_optimized}
\alias{search_multi_wildcard_optimized}
\title{Optimized multi-wildcard search}
\usage{
search_multi_wildcard_optimized(sequences, parts, n_sequences)
}
\description{
Optimized multi-wildcard search
}
\keyword{internal}
