% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_patterns.R
\name{search_pattern_optimized}
\alias{search_pattern_optimized}
\title{Optimized pattern search}
\usage{
search_pattern_optimized(
  sequences,
  pattern,
  n_sequences,
  n_states,
  test_significance,
  correction,
  alpha,
  fast_mode
)
}
\description{
Optimized pattern search
}
\keyword{internal}
