% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_patterns.R
\name{search_pattern_vectorized}
\alias{search_pattern_vectorized}
\title{Search Pattern (Vectorized)}
\usage{
search_pattern_vectorized(seq_data, pattern, verbose)
}
\description{
Search Pattern (Vectorized)
}
\keyword{internal}
