% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_rules.R
\name{summary.association_rules}
\alias{summary.association_rules}
\title{Summary Method for Association Rules}
\usage{
\method{summary}{association_rules}(object, ...)
}
\arguments{
\item{object}{Association rules object}

\item{...}{Additional arguments}
}
\description{
Summary Method for Association Rules
}
