% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_rules.R
\name{summary.bootstrapped_association_rules}
\alias{summary.bootstrapped_association_rules}
\title{Summary Method for Bootstrapped Association Rules}
\usage{
\method{summary}{bootstrapped_association_rules}(object, ...)
}
\arguments{
\item{object}{Bootstrapped association rules object}

\item{...}{Additional arguments}
}
\description{
Summary Method for Bootstrapped Association Rules
}
