% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_comparison.R
\name{summary.compare_sequences}
\alias{summary.compare_sequences}
\title{Summary Method for compare_sequences Objects}
\usage{
\method{summary}{compare_sequences}(object, ...)
}
\arguments{
\item{object}{A compare_sequences object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary Method for compare_sequences Objects
}
