% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_comparison.R
\name{summary.compare_sequences_multi}
\alias{summary.compare_sequences_multi}
\title{Summary method for compare_sequences_multi objects}
\usage{
\method{summary}{compare_sequences_multi}(object, length = NULL, top_n = 10, ...)
}
\arguments{
\item{object}{compare_sequences_multi object}

\item{length}{which subsequence length to summarize (default: all)}

\item{top_n}{number of top patterns to show (default: 10)}

\item{...}{additional arguments}
}
\description{
Summary method for compare_sequences_multi objects
}
