% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_analysis.R
\name{summary.pattern_analysis_multi}
\alias{summary.pattern_analysis_multi}
\title{Summary method for pattern_analysis_multi objects}
\usage{
\method{summary}{pattern_analysis_multi}(object, measure = NULL, top_n = 10, ...)
}
\arguments{
\item{object}{pattern_analysis_multi object}

\item{measure}{which measure to summarize}

\item{top_n}{number of top patterns to show}

\item{...}{additional arguments}
}
\description{
Summary method for pattern_analysis_multi objects
}
