% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_patterns.R
\name{top_sequences}
\alias{top_sequences}
\title{Extract Top Sequences}
\usage{
top_sequences(x, top_n = 10)
}
\arguments{
\item{x}{An object of class \code{sequence_patterns} returned by \code{\link{discover_patterns}}}

\item{top_n}{Integer specifying number of top patterns to return. Default: 10}
}
\value{
A data frame containing the top N patterns
}
\description{
Returns the top N patterns ordered by frequency/count.
}
\examples{
\dontrun{
results <- discover_patterns(data)
top10 <- top_sequences(results, top_n = 10)
}

}
