% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_utils.R
\name{validate_sequence_params}
\alias{validate_sequence_params}
\title{Validate Sequence Analysis Parameters}
\usage{
validate_sequence_params(params)
}
\arguments{
\item{params}{Named list of parameters to validate}
}
\description{
Validates common parameters for sequence analysis functions.
}
\keyword{internal}
