% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_viz.R
\name{plot.association_rules}
\alias{plot.association_rules}
\title{Plot Method for Association Rules}
\usage{
\method{plot}{association_rules}(x, type = "scatter", ...)
}
\arguments{
\item{x}{Association rules object}

\item{type}{Type of plot ("scatter", "network", "frequency", "quality", "matrix")}

\item{...}{Additional arguments passed to specific plot functions}
}
\description{
Generic plot method for association rules objects.
}
