% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_comparison.R
\name{plot.compare_sequences}
\alias{plot.compare_sequences}
\title{Plot Method for compare_sequences Objects}
\usage{
\method{plot}{compare_sequences}(x, ...)
}
\arguments{
\item{x}{A compare_sequences object}

\item{...}{Additional arguments (unused)}
}
\description{
Plot Method for compare_sequences Objects
}
